Use catalog type identifier width in binary binding plan

+++ crates/msduck-sql/src/projection/collation_validation.rs
@@ -94,7 +94,7 @@ enum BinaryDirection {
     ToUnicode {
         style: Box<Expr>,
         trying: bool,
-        source_id: i32,
+        source_id: u8,
     },
 }
 struct BinaryInput {
